+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Jetforce Changelog
 
+### Unreleased
+
+- Files with the ".gemini" extension are now recognized as text/gemini.
+- The default index file has been changed from ".gemini" to "index.gmi".
+- Added a new --index-file flag that can be used to customize the name of the
+  gemini file that will be served when a directory is requested.
+
 ### v0.0.6 (2019-08-22)
 
 - Significant refactoring of the base application interface.